// src/graphql/generated/getGamesForPlayer.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: getGamesForPlayer
// ====================================================

export interface getGamesForPlayer_getGamesForPlayer {
  __typename: "Player";
  ID: string | null;
  name: string | null;
  score: number | null;
  playerID: string | null;
}

export interface getGamesForPlayer {
  getGamesForPlayer: getGamesForPlayer_getGamesForPlayer[] | null;
}